merge branche multisite There is some consequence on existing coop application. The migration process is the following: - configure at least one Site object - update coop-geo - manage.py schemamigration coop_local --auto - manage.py migrate coop_local - manage.py migrate coop_geo - manage.py migrate multisite - in the shell, run the following code to add the default site to all objects import coop from django.db import models site = Site.objects.get_current() for m in models.get_models():     if coop.models.URIModel in m.__mro__:         for o in m.objects_manager.all():             o.sites.add(site)             o.save()
On URIModels the 'objects' manager is now dedicate to the objects belowing in the current site.
The 'objects_manager' correspond to the manager for all the objects without taking care of the site

If you don't need multisite possibility, there is nothing more to do.... else
you can have to switch settings COOP_USE_SITES to True, and thus in the admin you can chose available site for instance.