Generalization of the organization/views/geoJson method. First the method has been moved to coop/views/geoJson. The aim of the views does not change : it is to provide geoJson features according to filters (filters are documented in the code itself).
Now the method takes a model as parameter and works for all models
possibly involved in a generic relation with a Located ou AreaLink model.
It is the case for Organization, Project and Event model.

The urls 'org/geojson' has also been remove and replaced by
'geojson/organization'.
The url '/geojson/' (used only by AllianceProvence) has been kept for
compatibilty, it will be removed soon.

In order to write such a general views, we have change some fields name:
- event_type in Event is now category
- active is now a field for every StaticURIModel subclasses.
So migration for the coop_local application is necessary.

Basic methods to return GeoJson features have moved to coop-geo application.
So coop-geo have to be updated to.